Implement output and reporting methods, rate limit
The previously documented, but unimplemented, methods of output and
reporting have now been implemented.

We also add the ability to rate-limit the output, and change the
reporting interval (as the default of 10,000 is not always desired for
different output rates).

We also update the README with some examples, and apply a few
corrections to the text.