Rework the report interval by MB of logs
We now emit reports of the statistics when we have crossed a threshold
of a number of MBs instead of by time.  This avoids having to use timers
or to poll checking the time before or after each log entry
generated. However, it does mean that when using a certain rate-limit,
the user will likely want to adjust the reporting interval accordingly
from the default of 10 MBs.