Add initial driver script; report off stats
A very rudimentary load driver script is added with this commit to show
collection and verification using /dev/null, an on-disk file, and via a
pipe.  Such configurations will help us determine what is possible on a
given without involving a logging system.

We have enhanced the verify-loader command to report its metrics
whenever it comes across the stats emitted by the loader.  This is very
useful when the loader embeds stats within the payload, allowing the
verification to see how well it is keeping up as well (records time
difference assuming loader and verify-loader share same timezone).