Make verify-loader output easier to parse
This commit attempts to make it easier to parse the statistics generated
by the verify-loader program.

All numbers and identifiers are separated by whitespace.  We added skips
and duplicates to the reports, and stopped emitting the loader embedded
statistics by default.  The stats emitted from verify-loader are
prefixed with `verify-loader - `.