Speed improvements - mainly from using pypy3
We have updated the `loader` to use `pypy3` by default, and updated the
container image (adding a `Makefile` for building).

The `verify-loader` now reads using a byte-stream instead of text stream
to ensure we ready from pipes with the pipe size, and to avoid unnecessary
`UTF-8` encodings to process what is read.

We also optimize the container image to remove intermediate installation
files.

We also applied `flake8` to all the code changes.